requirements.txt
fastapi
uvicorn
pillow
httpx
apscheduler
python-dotenv
